add temp message creating for fast sending

+++ app/Http/Controllers/Api/ChatController.php
@@ -53,6 +53,8 @@ class ChatController extends Controller
 
     public function destroy(Chat $chat)
     {
+        $this->authorize('tune', $chat);
+
         $chat->delete();
     }
